+++ frontend/CHANGELOG.md
@@ -20,4 +20,9 @@
 - Customer datasheet
 - User is able to modify and delete Customer
 
- 
+ ## [v0.3.0] - 2025-05-05
+### WIP Version
+- Errorhandling on customer delete/modify
+- Unique telefonszam and szemelyiszam
+
+